+++ packages/address-extractor/tests/extractor/ergoUtxoExtractor.spec.ts
@@ -1,7 +1,6 @@
 import { DataSource } from 'typeorm';
 import * as ergoLib from 'ergo-lib-wasm-nodejs';
 import { ErgoNetworkType } from '@rosen-bridge/scanner-interfaces';
-import axios from '@rosen-bridge/rate-limited-axios';
 
 import { ErgoUTXOExtractor } from '../../lib';
 import { createDatabase } from './utils.mock';
@@ -11,13 +10,6 @@ let dataSource: DataSource;
 let extractor: ErgoUTXOExtractor;
 
 describe('AddressExtractor', () => {
-  beforeAll(() => {
-    axios.initConfigs({
-      apiLimitRateRangeAsSeconds: 10,
-      apiLimitRules: [],
-    });
-  });
-
   beforeEach(async () => {
     dataSource = await createDatabase();
     extractor = new ErgoUTXOExtractor(